tsql-core/src/test/java/com/xy/xsql/tsql/converter/functions/metadata/ObjectPropertyConverterTest.java
package com.xy.xsql.tsql.converter.functions.metadata;

import com.xy.xsql.block.core.printer.MetaBlockPrinter;
import com.xy.xsql.block.model.BlockMeta;
import org.junit.Assert;
import org.junit.Test;

import java.io.StringWriter;

import static org.junit.Assert.*;

/**
 * Created by xiaoyao9184 on 2018/6/1.
 */
public class ObjectPropertyConverterTest {

    @Test
    public void testMetaPrint() throws Exception {
        BlockMeta b = ObjectPropertyConverter.meta;

        StringWriter writer = MetaBlockPrinter.print(b);

        System.out.println(writer);
        Assert.assertEquals(writer.toString(),
                "<OBJECTPROPERTY> ::=\n" +
                        "OBJECTPROPERTY ( id , property )");
    }

}